Newsletter/requirements.txt
streamlit==1.30.0
requests==2.32.3
python-dotenv==1.0.1
openai==0.28.0
# google-generativeai==0.3.0
google-generativeai>=0.7.0